+++ bill/invoice.go
@@ -142,7 +142,14 @@ func (inv *Invoice) ValidateWithContext(ctx context.Context) error {
 			cal.DateNotZero(),
 		),
 		validation.Field(&inv.OperationDate),
-		validation.Field(&inv.ValueDate),
+		validation.Field(&inv.ValueDate,
+			// The value date indicates indicates when taxes become liable which is also indicated by the tax point
+			// so if the tax point is set, the value date should not be set to avoid confusion.
+			validation.When(
+				inv.Tax != nil && inv.Tax.Point != cbc.KeyEmpty,
+				validation.Empty.Error("value date cannot be set when tax point is set"),
+			),
+		),
 		validation.Field(&inv.Currency,
 			validation.Required,
 			exRule,
@@ -232,6 +239,10 @@ func partyHasTaxIDCode(party *org.Party) bool {
 // After inverting the invoice is recalculated and any differences will raise
 // an error.
 func (inv *Invoice) Invert() error {
+	if inv.HasTags(tax.TagBypass) {
+		return fmt.Errorf("cannot invert an invoice with tag bypass")
+	}
+
 	payable := inv.Totals.Payable.Invert()
 
 	for _, row := range inv.Lines {
@@ -254,8 +265,8 @@ func (inv *Invoice) Invert() error {
 			row.Amount = row.Amount.Invert()
 		}
 	}
-	inv.Totals = nil
 
+	inv.Totals = nil
 	if err := inv.Calculate(); err != nil {
 		return err
 	}