+++ bill/tax.go
@@ -24,15 +24,36 @@ type Tax struct {
 	// the rounding model used.
 	Rounding cbc.Key `json:"rounding,omitempty" jsonschema:"title=Rounding Model"`
 
+	// Point is a code that identifies the event which triggers the tax liability,
+	// such as invoice issuance, delivery of goods, or receipt of payment.
+	Point cbc.Key `json:"point,omitempty" jsonschema:"title=Point"`
+
 	// Additional extensions that are applied to the invoice as a whole as opposed to specific
 	// sections.
 	Ext tax.Extensions `json:"ext,omitempty" jsonschema:"title=Extensions"`
+
+	// Notes contains tax-related notes, typically used for exemption reasons
+	// or other tax-specific explanations associated with particular tax categories.
+	Notes []*tax.Note `json:"notes,omitempty" jsonschema:"title=Notes"`
 
 	// Any additional data that may be required for processing, but should never
 	// be relied upon by recipients.
 	Meta cbc.Meta `json:"meta,omitempty" jsonschema:"title=Meta"`
 
 	tags []cbc.Key
+}
+
+// MergeNotes adds a tax note to the tax object, automatically handling nil data,
+// and returning a new updated instance.
+func (t *Tax) MergeNotes(notes ...*tax.Note) *Tax {
+	if len(notes) == 0 {
+		return t
+	}
+	if t == nil {
+		t = new(Tax)
+	}
+	t.Notes = append(t.Notes, notes...)
+	return t
 }
 
 // MergeExtensions makes it easier to add extensions to the tax object
@@ -80,6 +101,7 @@ func (t *Tax) Normalize(normalizers tax.Normalizers) {
 		t.Rounding = tax.RoundingRuleCurrency
 	}
 	t.Ext = tax.CleanExtensions(t.Ext)
+	tax.Normalize(normalizers, t.Notes)
 	normalizers.Each(t)
 }
 
@@ -90,7 +112,11 @@ func (t *Tax) ValidateWithContext(ctx context.Context) error {
 		validation.Field(&t.Rounding,
 			cbc.InKeyDefs(tax.RoundingRules),
 		),
+		validation.Field(&t.Point,
+			cbc.InKeyDefs(tax.PointDefs),
+		),
 		validation.Field(&t.Ext),
+		validation.Field(&t.Notes),
 		validation.Field(&t.Meta),
 	)
 }
@@ -120,6 +146,16 @@ func (t Tax) JSONSchemaExtend(schema *jsonschema.Schema) {
 				Const:       r.Key.String(),
 				Title:       r.Name.String(),
 				Description: r.Desc.String(),
+			}
+		}
+	}
+	if p, ok := schema.Properties.Get("point"); ok {
+		p.OneOf = make([]*jsonschema.Schema, len(tax.PointDefs))
+		for i, w := range tax.PointDefs {
+			p.OneOf[i] = &jsonschema.Schema{
+				Const:       w.Key.String(),
+				Title:       w.Name.String(),
+				Description: w.Desc.String(),
 			}
 		}
 	}