+++ regimes/in/scenarios_test.go
@@ -70,7 +70,7 @@ func TestInvoiceDocumentScenarios(t *testing.T) {
 	i.SetTags(tax.TagSimplified)
 	i.Customer = nil
 	require.NoError(t, i.Calculate())
-	assert.Len(t, i.Notes, 1)
-	assert.Equal(t, i.Notes[0].Src, tax.TagSimplified)
-	assert.Equal(t, i.Notes[0].Text, "Simplified Tax Invoice")
+	require.NotNil(t, i.Tax)
+	assert.Len(t, i.Tax.Notes, 1)
+	assert.Equal(t, "Simplified Tax Invoice", i.Tax.Notes[0].Text)
 }